update to rust 1.81 and fix bug in --clean option of the build command
following error could occur on windows:
Error: Das System kann die angegebene Datei nicht finden. (os error 2)
error: process didn't exit successfully: `target\debug\neptungen.exe build --clean` (exit code: 1)

this happens if the output folder for the generated page did not exist.

+++ src/render.rs
@@ -42,7 +42,7 @@ pub fn build(path: &Path, conf: &Config, clean: bool) -> Result<()> {
         conf.title.as_deref().unwrap_or("unknown"),
     );
     let output_dir = PathBuf::from(conf.output_dir.as_ref().unwrap());
-    if clean {
+    if clean && output_dir.exists() {
         fs::remove_dir_all(&output_dir)?;
     }
     let nav_items = prepare_site_structure(path, output_dir.as_path(), conf);
